+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/models/events/ToolReservedEvent.java
@@ -1,5 +1,5 @@
 package com.veteroch4k.factory_service.models.events;
 
-public record ToolReservedEvent(long orderId) {
+public record ToolReservedEvent(int orderId) {
 
 }
