+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/repository/FactoryOrderRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FactoryOrderRepository extends CrudRepository<FactoryOrder, Long> {
+public interface FactoryOrderRepository extends CrudRepository<FactoryOrder, Integer> {
 
 }
