+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/repository/FactoryRepository.java
@@ -3,6 +3,6 @@ package com.veteroch4k.factory_service.repository;
 import com.veteroch4k.factory_service.models.Factory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FactoryRepository extends JpaRepository<Factory, Long> {
+public interface FactoryRepository extends JpaRepository<Factory, Integer> {
 
 }
